# BackendApplication/src/main/resources/openapi/openapi.yaml
openapi: 3.0.1
info:
  title: WrocłArts API
  description: REST API documentation for the WrocłArts system managing artistic activities in Wrocław.
  termsOfService: http://www.example.com/terms/
  contact:
    email: contact@maxime-luce.fr
  license:
    name: Apache-2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
...
servers:
  - url: http://localhost:8080/WroclArts
    description: Local Spring Boot Development Server
...
paths:
...
  /organizations:
    get:
      tags:
        - organization
      summary: find all organizations
      operationId: findAllOrganizations
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organizations'
            application/xml:
              schema:
                $ref: '#/components/schemas/Organizations'
      deprecated: false
      security:
        - wroclarts_auth:
            - read:organizations
    post:
      tags:
        - organization
      summary: register a new organization
      operationId: addOrganization
      requestBody:
        description: Organization object to register
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Organization'
          application/xml:
            schema:
              $ref: '#/components/schemas/Organization'
        required: true
      responses:
        200:
          description: Organization registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
            application/xml:
              schema:
                $ref: '#/components/schemas/Organization'
        405:
          description: Validation exception
          content: {}
      security:
        - wroclarts_auth:
            - write:organizations
      x-codegen-request-body-name: organization
  /organizations/{orgId}:
    get:
      tags:
        - organization
      summary: find an organization by orgId
      operationId: findOrganizationById
      parameters:
        - name: orgId
          in: path
          description: Numeric ID of the organization to get.
          required: true
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
            application/xml:
              schema:
                $ref: '#/components/schemas/Organization'
        400:
          description: Invalid orgId supplied
          content: { }
        404:
          description: Organization not found
          content: { }
      deprecated: false
      security:
        - wroclarts_auth:
            - read:organizations

    delete:
      tags:
        - organization
      summary: delete an organization by orgId
      operationId: deleteOrganizationById
      parameters:
        - name: orgId
          in: path
          description: Numeric ID of the organization to delete.
          required: true
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
            application/xml:
              schema:
                $ref: '#/components/schemas/Organization'
        400:
          description: Invalid orgId supplied
          content: { }
        404:
          description: Organization not found
          content: { }
      deprecated: false
      security:
        - wroclarts_auth:
            - write:organizations

    # --- LE NOUVEAU PUT EST ICI ---
    put:
      tags:
        - organization
      summary: update an existing organization profile
      operationId: updateOrganization
      parameters:
        - name: orgId
          in: path
          description: Numeric ID of the organization to update.
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Organization object to update
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Organization'
          application/xml:
            schema:
              $ref: '#/components/schemas/Organization'
        required: true
      responses:
        200:
          description: Organization updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
            application/xml:
              schema:
                $ref: '#/components/schemas/Organization'
        404:
          description: Organization not found
          content: { }
        405:
          description: Validation exception
          content: { }
      security:
        - wroclarts_auth:
            - write:organizations
      x-codegen-request-body-name: organization
  /products:
    get:
      tags:
        - product
      summary: find all products
      operationId: findAllProducts
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Products'
            application/xml:
              schema:
                $ref: '#/components/schemas/Products'
      deprecated: false
      security:
        - wroclarts_auth:
            - read:products
    post:
      tags:
        - product
      summary: add a new product to shop
      operationId: addProduct
      requestBody:
        description: Product object to add
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
          application/xml:
            schema:
              $ref: '#/components/schemas/Product'
        required: true
      responses:
        200:
          description: Product added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
            application/xml:
              schema:
                $ref: '#/components/schemas/Product'
        405:
          description: Validation exception
          content: {}
      security:
        - wroclarts_auth:
            - write:products
      x-codegen-request-body-name: product
  /products/{productId}:
    get:
      tags:
        - product
      summary: find a product by productId
      operationId: findProductById
      parameters:
        - name: productId
          in: path
          description: Numeric ID of the product.
          required: true
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
            application/xml:
              schema:
                $ref: '#/components/schemas/Product'
        400:
          description: Invalid productId supplied
          content: {}
        404:
          description: Product not found
          content: {}
      deprecated: false
      security:
        - wroclarts_auth:
            - read:products
    put:
      tags:
        - product
      summary: update an existing product
      operationId: updateProduct
      parameters:
        - name: productId
          in: path
          description: Numeric ID of the product.
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Product object to update
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
          application/xml:
            schema:
              $ref: '#/components/schemas/Product'
        required: true
      responses:
        200:
          description: Product updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
            application/xml:
              schema:
                $ref: '#/components/schemas/Product'
        404:
          description: Product not found
          content: {}
        405:
          description: Validation exception
          content: {}
      security:
        - wroclarts_auth:
            - write:products
      x-codegen-request-body-name: product
  /orders:
    get:
      tags:
        - order
      summary: find all orders
      operationId: findAllOrders
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Orders'
            application/xml:
              schema:
                $ref: '#/components/schemas/Orders'
      deprecated: false
      security:
        - wroclarts_auth:
            - read:orders
    post:
      tags:
        - order
      summary: Create a new order (checkout cart)
      operationId: createOrder
      requestBody:
        description: Order creation details
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Order'
          application/xml:
            schema:
              $ref: '#/components/schemas/Order'
        required: true
      responses:
        200:
          description: Order placed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
            application/xml:
              schema:
                $ref: '#/components/schemas/Order'
        400:
          description: Item out of stock or payment failed
          content: {}
      security:
        - wroclarts_auth:
            - write:orders
      x-codegen-request-body-name: order
  /orders/{orderId}:
    get:
      tags:
        - order
      summary: find an order by orderId
      operationId: findOrderById
      parameters:
        - name: orderId
          in: path
          description: Numeric ID of the order.
          required: true
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
            application/xml:
              schema:
                $ref: '#/components/schemas/Order'
        400:
          description: Invalid orderId supplied
          content: {}
        404:
          description: Order not found
          content: {}
      deprecated: false
      security:
        - wroclarts_auth:
            - read:orders
    put:
      tags:
        - order
      summary: update an existing order (e.g., status to Shipped)
      operationId: updateOrder
      parameters:
        - name: orderId
          in: path
          description: Numeric ID of the order.
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: Order object to update
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Order'
          application/xml:
            schema:
              $ref: '#/components/schemas/Order'
        required: true
      responses:
        200:
          description: Order updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
            application/xml:
              schema:
                $ref: '#/components/schemas/Order'
        404:
          description: Order not found
          content: {}
        405:
          description: Validation exception
          content: {}
      security:
        - wroclarts_auth:
            - write:orders
      x-codegen-request-body-name: order
...
components:
  schemas:
...
    Organization:
      required:
        - email
        - krsNumber
        - nameOrga
      type: object
      xml:
        name: Organization
      properties:
        orgId:
          type: integer
          format: int64
        email:
          type: string
          example: contact@ceramicstudio.pl
        nameOrga:
          type: string
          example: Ceramic Studio Wrocław
        krsNumber:
          type: string
          example: "0000123456"
        subscriptionDate:
          type: string
          format: date
          example: "2026-04-10"
        status:
          type: string
          example: PendingVerification
    Organizations:
      type: array
      items:
        $ref: '#/components/schemas/Organization'
      xml:
        name: Organizations
        wrapped: true
      example: |-
        <Organizations>
          <Organization>
            <orgId>1</orgId>
            <email>contact@ceramicstudio.pl</email>
            <krsNumber>0000123456</krsNumber>
            <status>PendingVerification</status>
          </Organization>
        </Organizations>
    Product:
      required:
        - name
        - price
      type: object
      xml:
        name: Product
      properties:
        productId:
          type: integer
          format: int64
        name:
          type: string
          example: Ceramic Mug
        price:
          type: number
          format: double
          example: 45.0
        stockQuantity:
          type: integer
          format: int32
          example: 10
        status:
          type: string
          example: Active
    Products:
      type: array
      items:
        $ref: '#/components/schemas/Product'
      xml:
        name: Products
        wrapped: true
      example: |-
        <Products>
          <Product>
            <productId>1</productId>
            <name>Ceramic Mug</name>
            <price>45.0</price>
            <stockQuantity>10</stockQuantity>
            <status>Active</status>
          </Product>
        </Products>
    Order:
      required:
        - status
        - deliveryMethod
      type: object
      xml:
        name: Order
      properties:
        orderId:
          type: integer
          format: int64
        status:
          type: string
          example: Shipped
        deliveryMethod:
          type: string
          example: Postal
        trackingId:
          type: string
          example: PL987654321
    Orders:
      type: array
      items:
        $ref: '#/components/schemas/Order'
      xml:
        name: Orders
        wrapped: true
      example: |-
        <Orders>
          <Order>
            <orderId>1</orderId>
            <status>Shipped</status>
            <deliveryMethod>Postal</deliveryMethod>
            <trackingId>PL987654321</trackingId>
          </Order>
        </Orders>
...
  securitySchemes:
    wroclarts_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: http://localhost/oauth/dialog
          scopes:
            write:events: modify events records in your account
            read:events: read your events records
            write:organizations: modify organization records
            read:organizations: read organization records
            write:products: modify shop products
            read:products: read shop products
            write:orders: modify order status
            read:orders: read order status
            admin:platform: exclusive platform administrator actions (moderation, verification)